Refactored into using router.route(), moved error handling file into middlewares folder and renamed all route files to match standards
