+++ routes/articles.routes.js
@@ -10,14 +10,13 @@ const {
 
 const articlesRouter = require('express').Router();
 
-articlesRouter.get('/', getArticles);
+articlesRouter.route('/').get(getArticles);
 
-articlesRouter.get('/:article_id', getArticleById);
+articlesRouter.route('/:article_id').get(getArticleById).patch(updateArticle);
 
-articlesRouter.get('/:article_id/comments', getCommentsForArticle);
-
-articlesRouter.post('/:article_id/comments', addCommentForArticle);
-
-articlesRouter.patch('/:article_id', updateArticle);
+articlesRouter
+  .route('/:article_id/comments')
+  .get(getCommentsForArticle)
+  .post(addCommentForArticle);
 
 module.exports = articlesRouter;
