+++ routes/topics.routes.js
@@ -2,6 +2,6 @@ const { getTopics } = require('../controllers/topics.controller');
 
 const topicsRouter = require('express').Router();
 
-topicsRouter.get('/', getTopics);
+topicsRouter.route('/').get(getTopics);
 
 module.exports = topicsRouter;
